test(telemetry): move agentless URL test into client_test.go
Consolidate the DD_SITE-derived agentless URL test alongside the
existing defaultConfig tests instead of a separate file.
